mini project/requirements.txt
streamlit
pandas
numpy
scikit-learn
scipy
torch
torchvision
pillow
joblib